+++ src/symbol.c
@@ -3,7 +3,8 @@
 #include "symbol.h"
 #include "util.h"
 
-Symbol* newSymbol(const char* name, u64 precedence, bool reserved) {
+
+Symbol* newSymbol(const char* name, u32 precedence, bool reserved) {
     Symbol* symbol = pMalloc(sizeof (Symbol));
 
     symbol->name = name;
